moved javascript classes and hardcoded datasources into their own folder
