+++ src/renderer/classes/measurements.js
