+++ src/renderer/classes/shoppingList.js
